+++ internal/providerfwk/resource_security_zone_book_address.go
@@ -78,7 +78,7 @@ func (rsc *securityZoneBookAddress) Schema(
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
 				Computed:    true,
-				Description: "An identifier for the resource with format `<zone>_-_<name>`.",
+				Description: "An identifier for the resource with format `<zone>" + junos.IDSeparator + "<name>`.",
 				PlanModifiers: []planmodifier.String{
 					stringplanmodifier.UseStateForUnknown(),
 				},