+++ internal/tfvalidator/string_net_test.go
@@ -37,6 +37,10 @@ func TestStringIPAddress(t *testing.T) {
 			val:         types.StringValue("2001:2::1"),
 			expectError: false,
 		},
+		"empty": {
+			val:         types.StringValue(""),
+			expectError: true,
+		},
 		"invalid": {
 			val:         types.StringValue("292.0.2.1"),
 			expectError: true,
@@ -125,6 +129,10 @@ func TestStringCIDR(t *testing.T) {
 		"valid": {
 			val:         types.StringValue("192.0.2.1/24"),
 			expectError: false,
+		},
+		"empty": {
+			val:         types.StringValue(""),
+			expectError: true,
 		},
 		"invalid": {
 			val:         types.StringValue("192.0.2.1"),
@@ -219,6 +227,10 @@ func TestStringCIDRNetwork(t *testing.T) {
 		"valid": {
 			val:         types.StringValue("192.0.2.0/24"),
 			expectError: false,
+		},
+		"empty": {
+			val:         types.StringValue(""),
+			expectError: true,
 		},
 		"invalid": {
 			val:         types.StringValue("192.0.2.1/24"),